wip: drag map
the current solution is far from ideal. We use DOM mouse events to
calculate the offset. However, the coornidates returned by Mouse
is in the browser space, not in the svg user space. That's why
we have to hard code a "0.005" to "translate".

A better choice may be native svg events which can give us coordinates
in the user space.

That said, this is just a experiment. We'll continue later.